% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chesschess-package.R
\docType{package}
\name{chesschess-package}
\alias{chesschess}
\alias{chesschess-package}
\title{chesschess: Chess Engine}
\description{
..
}
\author{
\strong{Maintainer}: Hugh Parsonage \email{hugh.parsonage@gmail.com}

}
\keyword{internal}
