% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enpassant.R
\name{enpassant}
\alias{enpassant}
\title{Enpassant}
\usage{
enpassant(
  x,
  y,
  start = 0L,
  white_to_move = TRUE,
  last_move = ifelse(white_to_move, y[1], x[1])
)
}
\arguments{
\item{x, y, start, white_to_move, last_move}{See \code{\link{is_checkmate}}.}
}
\value{
An integer: 0 if no enpassant possible; otherwise the column from which it is
possible, negative if the capture is to the left, positive if to the right.
}
\description{
Enpassant
}
