% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_checkmate.R
\name{is_checkmate}
\alias{is_checkmate}
\title{Is this checkmate?}
\usage{
is_checkmate(x, y, start = 0L, white_to_move = TRUE)
}
\arguments{
\item{x}{Positions of white, given as a character vector}

\item{y}{Positions of black}

\item{start}{Either 0 meaning a blank board, plus the positions given by \code{x, y}
or \code{1} meaning the starting positions plus \code{x, y}}

\item{white_to_move}{Is white to move (i.e. is white in checkmate?)}
}
\description{
Is this checkmate?
}
